Move contents of entities.ts into models.database.ts.
